Get Examiner Name from Office Actions

Extract the examiner's name from office action body text.

extract_examiner_name(office_action, collapse = TRUE)

Arguments

office_action

A list of office action, as retrieved with download_office_actions; each entry should be a list with entries for bodyText (the office action text) and examinerEmplyeeNumber.

collapse

Logical; if FALSE, will return a row for every entered office action.

Value

A character vector with the Examiner's employee number and two possible names: one extracted from the language (intext) and another from the sign-off (signed). The name will be NA if there is no examiner (such as if a non-examiner submitted the action), or if the format is not as expected. If office_action is a list of multiple office actions, these will make up rows of a data.frame, which will be as many as the length of office_actions if collapse is FALSE, or the length of unique IDs.

Examples

if (FALSE) {
# retrieve the office actions associated with a particular application
office_actions <- download_office_actions("patentApplicationNumber:13877637")

# get the name of the examiner associated with the first action:
extract_examiner_name(office_actions[[1]])
}