Download Patents or Patent Applications

Downloads from the U.S. Pre-Grant Publications (US-PGPUB), Patents (USPAT), and/or Optical Character Recognition (USOCR) databases. See the web app to search for and view documents from the same source.

uspto_download(guid, outDir = tempdir(), type = "US-PGPUB",
  cores = detectCores() - 2, compress = TRUE, load = TRUE,
  verbose = FALSE)

Arguments

guid

Vector of document numbers (e.g., US-20220230294-A1). Can also be a data.frame of results, as returned from uspto_search; must have a guid column containing document numbers, and a type column with the source database abbreviation (e.g., US-PGPUB).

outDir

Path to a directory in which to save each individual document. If not specified, these are saved to a temporary directory.

type

A vector the same length as guid (repeated as necessary), indicating the source database of the guid. Defaults to US-PGPUB for every ID.

cores

Number of CPU cores to split requests across.

compress

Logical; if FALSE, will not xz-compress each file.

load

Logical; if FALSE, will not return content, or load existing files.

verbose

Logical; if FALSE, does not print status messages.

Value

A data.frame with a row for each document.

See also

Examples

if (FALSE) {
# start with a search
results <- uspto_search("G01W.CPCL.", "US-PGPUB")

# then download those results
applications <- uspto_download(results)
}