Search for Patents or Patent Applications

Searches the U.S. Pre-Grant Publications (US-PGPUB), Patents (USPAT), and/or Optical Character Recognition (USOCR) databases. See the web app for an interactive search.

uspto_search(query, databases = c("US-PGPUB", "USPAT", "USOCR"),
  outFile = NULL, start = 0, limit = FALSE, sort = "date_publ desc",
  english_only = TRUE, spellCheck = FALSE, plurals = TRUE,
  britishEquivalents = TRUE, verbose = FALSE)

Arguments

query

A query string; for example "noodle", "photographic AND noodle", "wet adj4 silicon", or "G01W.CPCL.". See the Patent Public Search training materials.

databases

A character vector specifying which databases to search in; a selection from c("US-PGPUB", "USPAT", "USOCR").

outFile

Name of a .csv file to save results to. If this file exists, it will be loaded instead of searching.

start

Result of the search to start from.

limit

Maximum number of results to return; defaults to all results.

sort

How to sort results; defaults to publication date ("date_publ desc").

english_only

Logical; if FALSE, will return patent and/or applications of any language.

spellCheck

Logical; if TRUE, will spellcheck query terms.

plurals

Logical; if TRUE, will look for plural and singular forms of query terms.

britishEquivalents

Logical; if TRUE, will look for British and American forms of query terms.

verbose

Logical; if FALSE, does not print status messages.

Value

A data.frame of results, if any were found; otherwise NULL.

Examples

if (FALSE) {
# search for applications with a meteorology classification:
# https://www.uspto.gov/web/patents/classification/cpc/html/cpc-G01W.html
results <- uspto_search("G01W.CPCL.", "US-PGPUB")

# search by date and keyword:
results <- uspto_search("20020801.pd. AND motorcycle")
}